Refactor code structure and adjust URLs
Various components have been grouped into 'components' and 'pages' directories to establish a more logical structure and improve maintainability. In addition, logout URL and home link in side-bar has been revised to redirect to accurate pages. These changes require modifying import paths and URLs in many places in the code, but will make it clearer and more understandable.
