+++ src/components/store/storeList/StoreList.jsx
@@ -1,3 +1,4 @@
+// eslint-disable-next-line no-unused-vars
 import React, { useState } from 'react';
 
 var path = 'tripleh.jpg';
