+++ src/pages/mainPage/Slider.jsx
